backend/requirements.txt
django==3.2
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
django-cors-headers==3.7.0
django-filter==2.4.0
djoser==2.1.0
psycopg2-binary==2.9.2